split database file to database plugin and database api;  removed allowWrite logic and unused checkDB logic
