+++ osquery/database/database.cpp
@@ -12,6 +12,7 @@
 #include <boost/property_tree/json_parser.hpp>
 
 #include <osquery/database.h>
+#include <osquery/database_plugin.h>
 #include <osquery/flags.h>
 #include <osquery/logger.h>
 #include <osquery/registry.h>
@@ -23,8 +24,13 @@ namespace rj = rapidjson;
 
 namespace osquery {
 
+/// An internal status protecting database access.
+static std::atomic<bool> kDBInitialized;
+
 /// Generate a specific-use registry for database access abstraction.
 CREATE_REGISTRY(DatabasePlugin, "database");
+
+FLAG(bool, disable_database, false, "Disable the persistent RocksDB storage");
 
 CLI_FLAG(bool, database_dump, false, "Dump the contents of the backing store");
 
@@ -33,8 +39,6 @@ CLI_FLAG(string,
          OSQUERY_DB_HOME "/osquery.db",
          "If using a disk-based backing store, specify a path");
 FLAG_ALIAS(std::string, db_path, database_path);
-
-FLAG(bool, disable_database, false, "Disable the persistent RocksDB storage");
 
 const std::string kInternalDatabase = "rocksdb";
 const std::string kPersistentSettings = "configurations";
@@ -51,11 +55,6 @@ const std::string kDatabaseResultsVersion = "1";
 const std::vector<std::string> kDomains = {
     kPersistentSettings, kQueries, kEvents, kLogs, kCarves};
 
-std::atomic<bool> DatabasePlugin::kDBAllowOpen(false);
-std::atomic<bool> DatabasePlugin::kDBRequireWrite(false);
-std::atomic<bool> DatabasePlugin::kDBInitialized(false);
-std::atomic<bool> DatabasePlugin::kDBChecking(false);
-
 /**
  * @brief A reader/writer mutex protecting database resets.
  *
@@ -63,120 +62,6 @@ std::atomic<bool> DatabasePlugin::kDBChecking(false);
  * database plugin APIs.
  */
 Mutex kDatabaseReset;
-
-Status DatabasePlugin::initPlugin() {
-  // Initialize the database plugin using the flag.
-  auto plugin = (FLAGS_disable_database) ? "ephemeral" : kInternalDatabase;
-  auto status = RegistryFactory::get().setActive("database", plugin);
-  if (!status.ok()) {
-    // If the database did not setUp override the active plugin.
-    RegistryFactory::get().setActive("database", "ephemeral");
-  }
-
-  kDBInitialized = true;
-  return status;
-}
-
-void DatabasePlugin::shutdown() {
-  auto datbase_registry = RegistryFactory::get().registry("database");
-  for (auto& plugin : RegistryFactory::get().names("database")) {
-    datbase_registry->remove(plugin);
-  }
-}
-
-Status DatabasePlugin::reset() {
-  // Keep this simple, scope the critical section to the broader methods.
-  tearDown();
-  return setUp();
-}
-
-bool DatabasePlugin::checkDB() {
-  kDBChecking = true;
-  bool result = true;
-  try {
-    auto status = setUp();
-    if (kDBRequireWrite && read_only_) {
-      result = false;
-    }
-    tearDown();
-    result = status.ok();
-  } catch (const std::exception& e) {
-    VLOG(1) << "Database plugin check failed: " << e.what();
-    result = false;
-  }
-  kDBChecking = false;
-  return result;
-}
-
-Status DatabasePlugin::scan(const std::string& domain,
-                            std::vector<std::string>& results,
-                            const std::string& prefix,
-                            size_t max) const {
-  return Status(0, "Not used");
-}
-
-Status DatabasePlugin::call(const PluginRequest& request,
-                            PluginResponse& response) {
-  if (request.count("action") == 0) {
-    return Status(1, "Database plugin must include a request action");
-  }
-
-  // Get a domain/key, which are used for most database plugin actions.
-  auto domain = (request.count("domain") > 0) ? request.at("domain") : "";
-  auto key = (request.count("key") > 0) ? request.at("key") : "";
-
-  if (request.at("action") == "reset") {
-    WriteLock lock(kDatabaseReset);
-    DatabasePlugin::kDBInitialized = false;
-    // Prevent RocksDB reentrancy by logger plugins during plugin setup.
-    VLOG(1) << "Resetting the database plugin: " << getName();
-    auto status = this->reset();
-    if (!status.ok()) {
-      // The active database could not be reset, fallback to an ephemeral.
-      Registry::get().setActive("database", "ephemeral");
-      LOG(WARNING) << "Unable to reset database plugin: " << getName();
-    }
-    DatabasePlugin::kDBInitialized = true;
-    return status;
-  }
-
-  // Switch over the possible database plugin actions.
-  ReadLock lock(kDatabaseReset);
-  if (request.at("action") == "get") {
-    std::string value;
-    auto status = this->get(domain, key, value);
-    response.push_back({{"v", value}});
-    return status;
-  } else if (request.at("action") == "put") {
-    if (request.count("value") == 0) {
-      return Status(1, "Database plugin put action requires a value");
-    }
-    return this->put(domain, key, request.at("value"));
-  } else if (request.at("action") == "remove") {
-    return this->remove(domain, key);
-  } else if (request.at("action") == "remove_range") {
-    auto key_high = (request.count("high") > 0) ? request.at("key_high") : "";
-    if (!key_high.empty() && !key.empty()) {
-      return this->removeRange(domain, key, key_high);
-    }
-    return Status(1, "Missing range");
-  } else if (request.at("action") == "scan") {
-    // Accumulate scanned keys into a vector.
-    std::vector<std::string> keys;
-    // Optionally allow the caller to request a max number of keys.
-    size_t max = 0;
-    if (request.count("max") > 0) {
-      max = std::stoul(request.at("max"));
-    }
-    auto status = this->scan(domain, keys, request.at("prefix"), max);
-    for (const auto& k : keys) {
-      response.push_back({{"k", k}});
-    }
-    return status;
-  }
-
-  return Status(1, "Unknown database plugin action");
-}
 
 static inline std::shared_ptr<DatabasePlugin> getDatabasePlugin() {
   auto& rf = RegistryFactory::get();
@@ -186,6 +71,10 @@ static inline std::shared_ptr<DatabasePlugin> getDatabasePlugin() {
 
   auto plugin = rf.plugin("database", rf.getActive("database"));
   return std::dynamic_pointer_cast<DatabasePlugin>(plugin);
+}
+
+bool isDatabaseInitilized() {
+  return kDBInitialized;
 }
 
 Status getDatabaseValue(const std::string& domain,
@@ -212,7 +101,7 @@ Status getDatabaseValue(const std::string& domain,
   }
 
   ReadLock lock(kDatabaseReset);
-  if (!DatabasePlugin::kDBInitialized) {
+  if (!kDBInitialized) {
     throw std::runtime_error("Cannot get database value: " + key);
   } else {
     auto plugin = getDatabasePlugin();
@@ -247,7 +136,7 @@ Status setDatabaseValue(const std::string& domain,
   }
 
   ReadLock lock(kDatabaseReset);
-  if (!DatabasePlugin::kDBInitialized) {
+  if (!kDBInitialized) {
     throw std::runtime_error("Cannot set database value: " + key);
   } else {
     auto plugin = getDatabasePlugin();
@@ -275,7 +164,7 @@ Status deleteDatabaseValue(const std::string& domain, const std::string& key) {
   }
 
   ReadLock lock(kDatabaseReset);
-  if (!DatabasePlugin::kDBInitialized) {
+  if (!kDBInitialized) {
     throw std::runtime_error("Cannot delete database value: " + key);
   } else {
     auto plugin = getDatabasePlugin();
@@ -301,7 +190,7 @@ Status deleteDatabaseRange(const std::string& domain,
   }
 
   ReadLock lock(kDatabaseReset);
-  if (!DatabasePlugin::kDBInitialized) {
+  if (!kDBInitialized) {
     throw std::runtime_error("Cannot delete database values: " + low + " - " +
                              high);
   } else {
@@ -314,6 +203,26 @@ Status scanDatabaseKeys(const std::string& domain,
                         std::vector<std::string>& keys,
                         size_t max) {
   return scanDatabaseKeys(domain, keys, "", max);
+}
+
+Status initializeDatabase() {
+  // Initialize the database plugin using the flag.
+  auto plugin = (FLAGS_disable_database) ? "ephemeral" : kInternalDatabase;
+  auto status = RegistryFactory::get().setActive("database", plugin);
+  if (!status.ok()) {
+    // If the database did not setUp override the active plugin.
+    RegistryFactory::get().setActive("database", "ephemeral");
+  }
+
+  kDBInitialized = true;
+  return status;
+}
+
+void shutdownDatabase() {
+  auto datbase_registry = RegistryFactory::get().registry("database");
+  for (auto& plugin : RegistryFactory::get().names("database")) {
+    datbase_registry->remove(plugin);
+  }
 }
 
 /// Get a list of keys for a given domain.
@@ -344,7 +253,7 @@ Status scanDatabaseKeys(const std::string& domain,
   }
 
   ReadLock lock(kDatabaseReset);
-  if (!DatabasePlugin::kDBInitialized) {
+  if (!kDBInitialized) {
     throw std::runtime_error("Cannot scan database values: " + prefix);
   } else {
     auto plugin = getDatabasePlugin();
@@ -353,8 +262,10 @@ Status scanDatabaseKeys(const std::string& domain,
 }
 
 void resetDatabase() {
+  kDBInitialized = false;
   PluginRequest request = {{"action", "reset"}};
   Registry::call("database", request);
+  kDBInitialized = true;
 }
 
 void dumpDatabase() {
