+++ osquery/main/main.cpp
@@ -77,7 +77,6 @@ int profile(int argc, char* argv[]) {
   }
 
   // Perform some duplication from Initializer with respect to database setup.
-  osquery::DatabasePlugin::setAllowOpen(true);
   osquery::RegistryFactory::get().setActive("database", "ephemeral");
 
   auto dbc = osquery::SQLiteDBManager::get();
